requirements.txt
asgiref==3.4.1
autopep8==1.6.0
black==21.12b0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cryptography==36.0.1
defusedxml==0.7.1
Django==3.2.10
django-debug-toolbar==3.2.4
django-debug-toolbar-template-profiler==2.0.2
gunicorn==20.1.0
idna==3.3
isort==5.10.1
mypy-extensions==0.4.3
oauthlib==3.2.0
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.3
typing-extensions==4.0.1
urllib3==1.26.8
wrapt==1.13.3
